backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.0
psycopg2-binary
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-Migrate==4.0.0
marshmallow==3.21.0
cloudinary==1.38.0
gunicorn==22.0.0 
Flask-Cors